requirements.txt
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
imagehash>=4.3.1
tqdm>=4.65.0
rich>=13.0.0
jinja2>=3.1.0
click>=8.1.0
